requirements.txt
python-dotenv
requests
google-generativeai
pandas
openpyxl
chromadb